+++ src/navigation/homeNavigator.tsx
@@ -41,6 +41,13 @@ export const HomeNavigator = ({ navigation }) => {
                     return { headerTitle: routeName }
                 }}
             />
+            <Stack.Screen
+                name={routes.CREATE_BLOCK}
+                component={CreateBlock}
+                options={({ route }) => {
+                    const routeName = getFocusedRouteNameFromRoute(route) ?? 'Create Block'
+                    return { headerTitle: routeName }
+                }} />
         </Stack.Navigator>
     )
 }
